requirements.txt
pandas
torch
torchvision
d2l
numpy
matplotlib
nuscenes-devkit
tqdm